接口自动化框架/requirements.txt
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.15.0
allure-pytest==2.9.43
allure-python-commons==2.9.43
appdirs==1.4.4
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
configobj==5.0.6
crcmod==1.7
cryptography==3.4.8
docopt==0.6.2
Faker==8.13.2
idna==3.2
iniconfig==1.1.1
ipython-genutils==0.2.0
jmespath==0.10.0
jsonpath==0.82
jsonschema==3.2.0
jupyter-core==4.8.1
Logbook==1.5.3
mypy-extensions==0.4.3
natsort==7.1.1
nbformat==5.1.3
oss2==2.15.0
packaging==21.0
pathspec==0.9.0
pep517==0.11.0
pigar==0.10.0
pip-tools==6.3.0
pipreqs==0.4.10
pluggy==1.0.0
py==1.10.0
pycparser==2.20
pycryptodome==3.10.1
pyotp==2.6.0
pyparsing==2.4.7
pyrsistent==0.18.0
pytest==6.2.5
pytest-assume==2.4.3
pytest-dependency==0.5.1
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.2
python-dateutil==2.8.2
regex==2021.8.28
requests==2.26.0
requests-toolbelt==0.9.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
traitlets==5.1.0
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
xlrd==2.0.1
yarg==0.1.9
